Backend/requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
uuid==1.30
psycopg2-binary==2.9.10
flask_cors==5.0.0
pytest==8.3.4
flask_migrate==4.0.7
SQLAlchemy==2.0.36 
python-dateutil==2.9
flask_cors==5.0.0
requests==2.32
PyJWT==2.10.1
python-dotenv==1.0.0